v.0.3.1
Added a 'beta-branch' folder as well as a 'live' folder.

The 'beta-branch' folder intended use is to upload experimental versions of the same mod that will (eventually) make it into the live version.

The 'live' folder is intended to be a mirror of what we got uploaded to Nexus.

This folder structure will hopefully make it easier for us to maintain the mod with updates as well as allowing us more freedom to experiment with new ideas etc. that might not make it into the final release.