v. 0.4-beta
Hi @xaratas and sorry for constantly making changes to the mod! Must be hell to try and keep up with the translations. I'll try and keep the changes more limited in the future, promise!

Made a lot of QoL improvements to the stylization of the various keywords as well as fixing a few smaller issues with specific keywords.

I'm sure that I both missed and broke something with this update so let me know if something stands out!

Also: perhaps we should change the main folder name from " 	enhanced_ui_affliction_inspiration" to "PoE2-EnhancedUserInterface" so that it corresponds to the new mod name?

See the change.log for a more detailed explaination of the changes I've made.